OCPBUGS-2319: Set NODAD flag on masquerade address
When there are multiple addresses on the primary interface for a node,
conntrack uses the first one by default. During upgrade that ends
up being the masquerade address because it was added last. This
causes connection problems because the masquerade address ends up in
a dadfailed state because we use the same address on all nodes. Since
we don't care about DAD on that address, let's disable it so the
address doesn't get into a bad state.

Merge-Conflict in go-controller/go.mod resolved by Douglas

Signed-off-by: Ben Nemec <bnemec@redhat.com>
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>